+++ member/src/pages/[id]/search.vue
@@ -7,33 +7,19 @@
     <MusicCard
       v-for="music in state.musics"
       v-else
+      :id="music.id"
       :key="music.id"
       :thumbnail="music.thumbnail"
       :name="music.name"
       :artists="music.artists"
       @click="onClickMusicCard(music)"
     />
-    <ModalContainer v-bind="modalProps">
-      <template #content>
-        <MusicRequestForm
-          :music-name="state.modal.name"
-          :artist-name="state.modal.artists"
-          :album-url="state.modal.thumbnail"
-          :text-area="state.modal.textArea"
-          :text-field="state.modal.textField"
-          @change-text-field="changeTextField"
-          @change-text-area="changeTextArea"
-          @on-click-submit-button="onClickSubmitButton"
-        />
-      </template>
-    </ModalContainer>
   </div>
 </template>
 <script setup lang="ts">
-  import { music, useRequestTimer } from '@/features'
-  import { RequestMusicInput, SearchMusicInput, SearchMusicResponse } from '@/features/music/api'
+  import { music } from '@/features'
+  import { SearchMusicInput, SearchMusicResponse } from '@/features/music/api'
   const route = useRoute()
-  const requestTimer = useRequestTimer()
 
   const musicInit: SearchMusicResponse = []
 
@@ -62,31 +48,6 @@
     state.isModalOpen = true
   }
 
-  const changeTextField = (value: string) => {
-    state.modal.textField = value
-  }
-
-  const changeTextArea = (value: string) => {
-    state.modal.textArea = value
-  }
-
-  const onClickOutside = () => {
-    state.isModalOpen = false
-  }
-  const onClickCloseButton = () => {
-    state.isModalOpen = false
-  }
-
-  const modalProps = reactive({
-    isOpen: toRef(state, 'isModalOpen'),
-    onClickOutside,
-    onClickCloseButton
-  })
-
-  const onClickSubmitButton = () => {
-    requestMusic()
-  }
-
   const fetchMusics = async () => {
     state.loading = true
     if (!state.query) { state.loading = false; return }
@@ -100,25 +61,6 @@
     state.musics = result.data.value
     state.loading = false
     state.beforeQuery = state.query.toString()
-  }
-
-  const requestMusic = async () => {
-    const roomId: string = String(route.params.id)
-    const requestMusicInput: RequestMusicInput = {
-      roomId,
-      musics: [state.modal.id],
-      radioName: state.modal.textField,
-      message: state.modal.textArea
-    }
-
-    const result = music.api.requestMusic(requestMusicInput)
-    await result.execute()
-    if (result.data.value) {
-      await navigateTo(`/${roomId}/requested`)
-      requestTimer.requestMusic()
-    } else {
-      alert(result.error.value)
-    }
   }
 
   onMounted(async () => {
